Move microstrain core files out of "common" subdir.
