+++ src/c/microstrain/connections/serial/CMakeLists.txt
@@ -16,7 +16,7 @@ endif()
 
 target_compile_features(${MICROSTRAIN_SERIAL_LIBRARY} PUBLIC c_std_11)
 
-target_link_libraries(${MICROSTRAIN_SERIAL_LIBRARY} PUBLIC ${MICROSTRAIN_COMMON_LIBRARY})
+target_link_libraries(${MICROSTRAIN_SERIAL_LIBRARY} PUBLIC ${MICROSTRAIN_LIBRARY})
 
 target_compile_definitions(${MICROSTRAIN_SERIAL_LIBRARY}
     PUBLIC "MICROSTRAIN_ENABLE_SERIAL=1"
