+++ src/c/microstrain/logging.h
