+++ src/c/microstrain/serialization.c
