+++ src/c/microstrain/serialization.h
