+++ src/c/mip/mip_packet.c
@@ -4,7 +4,7 @@
 
 #include "mip_descriptors.h"
 
-#include <microstrain/common/serialization.h>
+#include <microstrain/serialization.h>
 
 #include <string.h>
 #include <assert.h>
