+++ src/cpp/microstrain/connections/recording/CMakeLists.txt
@@ -17,7 +17,7 @@ endif()
 
 target_compile_features(${MICROSTRAIN_RECORDING_CONNECTION_LIBRARY} PUBLIC cxx_std_11)
 
-target_link_libraries(${MICROSTRAIN_RECORDING_CONNECTION_LIBRARY} PUBLIC ${MICROSTRAIN_COMMON_LIBRARY})
+target_link_libraries(${MICROSTRAIN_RECORDING_CONNECTION_LIBRARY} PUBLIC ${MICROSTRAIN_LIBRARY})
 
 target_compile_options(${MICROSTRAIN_RECORDING_CONNECTION_LIBRARY} PRIVATE ${MICROSTRAIN_PRIVATE_COMPILE_OPTIONS})
