+++ src/cpp/microstrain/extras/CMakeLists.txt
@@ -16,7 +16,7 @@ endif()
 
 target_compile_features(${MICROSTRAIN_EXTRAS_LIBRARY} PUBLIC cxx_std_11)
 
-target_link_libraries(${MICROSTRAIN_EXTRAS_LIBRARY} PUBLIC ${MICROSTRAIN_COMMON_LIBRARY})
+target_link_libraries(${MICROSTRAIN_EXTRAS_LIBRARY} PUBLIC ${MICROSTRAIN_LIBRARY})
 
 target_compile_definitions(${MICROSTRAIN_EXTRAS_LIBRARY}
     PUBLIC "MICROSTRAIN_ENABLE_EXTRAS=1"
