+++ src/cpp/microstrain/serialization.hpp
