+++ src/cpp/microstrain/serialization/serializer.hpp
