+++ src/cpp/microstrain/span.hpp
